Breakable walls and platformss
