+++ Assets/BasePrefabs/Breakable Wall.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 898bf2b2c9c9c3e42a85ad73805bc25f
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
